+++ src/extension.ts
@@ -1,12 +1,28 @@
 'use strict'
 
-import { languages, ExtensionContext, window } from 'vscode'
+import {
+    languages,
+    ExtensionContext,
+    window,
+    workspace
+} from 'vscode'
 import LinkProvider from './providers/linkProvider'
+import * as util from './util'
 
 let providers = []
 const debounce = require('lodash.debounce')
 
 export function activate(context: ExtensionContext) {
+    util.readConfig()
+
+    // config
+    workspace.onDidChangeConfiguration(async (e) => {
+        if (e.affectsConfiguration('laravel_goto_view')) {
+            util.readConfig()
+        }
+    })
+
+    // links
     setTimeout(() => {
         if (window.activeTextEditor) {
             initProvider()