+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - init
+
+## 0.0.4
+
+- add config for methods search
+- show underline under the key only instead of the whole line